+++ src/lib/variable_data.ts
@@ -1,7 +1,7 @@
 import type {Array_Element} from '@ryanatkn/belt/types.js';
 
-export type Size_Variant = Array_Element<typeof size_variants>;
-export const size_variants = [
+export type Size_Variant = Array_Element<typeof font_size_variants>;
+export const font_size_variants = [
 	'xs',
 	'sm',
 	'md',
@@ -17,7 +17,7 @@ export const size_variants = [
 	'xl9',
 ] as const;
 
-export const size_names = size_variants.map((s) => 'size_' + s);
+export const font_size_names = font_size_variants.map((s) => 'font_size_' + s);
 
 export type Text_Color_Variant = Array_Element<typeof text_color_variants>;
 export const text_color_variants = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10] as const;
@@ -69,21 +69,21 @@ export const shadow_variant_prefixes = [
 	'shadow_inset_bottom_',
 ] as const;
 
-export type Shadow_Size_Variant = Array_Element<typeof shadow_size_variants>;
-export const shadow_size_variants = ['xs', 'sm', 'md', 'lg', 'xl'] as const;
+export type Shadow_Size_Variant = Array_Element<typeof shadow_font_size_variants>;
+export const shadow_font_size_variants = ['xs', 'sm', 'md', 'lg', 'xl'] as const;
 
 export type Shadow_Alpha_Variant = Array_Element<typeof shadow_alpha_variants>;
 export const shadow_alpha_variants = [1, 2, 3, 4, 5] as const;
 
-// TODO maybe put this inline? factor out the pieces with `icon_size_variants`?
+// TODO maybe put this inline? factor out the pieces with `icon_font_size_variants`?
 export const icon_sizes = {
-	icon_size_xs: '18px',
-	icon_size_sm: '32px',
-	icon_size_md: '48px',
-	icon_size_lg: '80px',
-	icon_size_xl: '128px',
-	icon_size_xl2: '192px',
-	icon_size_xl3: '256px',
+	icon_font_size_xs: '18px',
+	icon_font_size_sm: '32px',
+	icon_font_size_md: '48px',
+	icon_font_size_lg: '80px',
+	icon_font_size_xl: '128px',
+	icon_font_size_xl2: '192px',
+	icon_font_size_xl3: '256px',
 };
 
 export type Color_Variant = Array_Element<typeof color_variants>;