+++ src/routes/+page.svelte
@@ -16,7 +16,7 @@
 <main class="box mb_xl7">
 	<section class="box">
 		<h1 class="mb_sm" style:margin-top={MAIN_HEADER_MARGIN_TOP}>moss</h1>
-		<Svg data={moss_logo} size="var(--icon_size_xl2)" />
+		<Svg data={moss_logo} size="var(--icon_font_size_xl2)" />
 		<div class="p_xl mt_lg border_radius_lg shadow_inset_bottom_xs">
 			<blockquote class="shadow_inset_top_xs">magical organic stylesheets 🌿</blockquote>
 		</div>