+++ CHANGELOG.md
@@ -73,7 +73,7 @@
 
 ### Patch Changes
 
-- add `--size` for `.size_*` classes and `--font_weight` for `.font_weight_*` classes ([#50](https://github.com/ryanatkn/moss/pull/50))
+- add `--font_size` for `.font_size_*` classes and `--font_weight` for `.font_weight_*` classes ([#50](https://github.com/ryanatkn/moss/pull/50))
 - add `.heading` for h1-6-like behavior ([#50](https://github.com/ryanatkn/moss/pull/50))
 
 ## 0.22.5